feat!: move runCommand into the default export
Export `runCommand`, `CommandArray`, `RunCommandOptions`, and
`RunCommandResult` from the main `@hugojosefson/log-fold` entry point.
Remove the separate `@hugojosefson/log-fold/run-command` subpath export.

BREAKING CHANGE: `import { runCommand } from "@hugojosefson/log-fold/run-command"`
no longer works. Use `import { runCommand } from "@hugojosefson/log-fold"`.

Fixes #20
