+++ mod.ts
@@ -14,6 +14,14 @@ export type {
   StreamPair,
 } from "./src/log-from-stream.ts";
 
+// Subprocess wrapper
+export { runCommand } from "./src/run-command.ts";
+export type {
+  CommandArray,
+  RunCommandOptions,
+  RunCommandResult,
+} from "./src/run-command.ts";
+
 // Types
 export type { SessionOptions } from "./src/session.ts";
 export type { Spinner, TaskOptions } from "./src/task-node.ts";
