+++ readme/example-subprocess-wrapper.ts
@@ -1,7 +1,6 @@
 #!/usr/bin/env -S deno run --allow-run=npm --allow-env
 
-import { logTask } from "../mod.ts";
-import { runCommand } from "../src/run-command.ts";
+import { logTask, runCommand } from "../mod.ts";
 
 await logTask("Run innocuous npm commands", async () => {
   await runCommand(["npm", "search", "typescript"]);
