+++ test/run-command.test.ts
@@ -1,8 +1,7 @@
 import { assertEquals, assertRejects } from "@std/assert";
 import { spawn } from "node:child_process";
-import { logFromStream, logTask } from "../mod.ts";
+import { logFromStream, logTask, runCommand } from "../mod.ts";
 import type { WriteStreamLike } from "../src/renderer/write-stream-like.ts";
-import { runCommand } from "../src/run-command.ts";
 
 /** Create a mock writable stream that collects output. */
 function mockStream(): WriteStreamLike & { lines: string[] } {
